Moved most contents of Enzo/particle into subdirectories called formation and feedback
