+++ src/Enzo/_enzo.hpp
@@ -215,9 +215,9 @@ extern "C" {
 #include "amr/EnzoRefineMass.hpp"
 
 // [order dependencies:]
-#include "particle/EnzoSinkParticle.hpp"
-#include "particle/EnzoBondiHoyleSinkParticle.hpp"
-#include "particle/EnzoFluxSinkParticle.hpp"
+#include "particle/formation/EnzoSinkParticle.hpp"
+#include "particle/formation/EnzoBondiHoyleSinkParticle.hpp"
+#include "particle/formation/EnzoFluxSinkParticle.hpp"
 
 // [order dependencies:]
 #include "utils/EnzoEFltArrayMap.hpp"
@@ -246,33 +246,33 @@ extern "C" {
 #include "hydro-mhd/EnzoBfieldMethod.hpp"
 #include "hydro-mhd/EnzoBfieldMethodCT.hpp"
 
-#include "particle/EnzoMethodAccretion.hpp"
+#include "particle/formation/EnzoMethodAccretion.hpp"
 #include "gravity/EnzoMethodBackgroundAcceleration.hpp"
-#include "particle/EnzoMethodBondiHoyleAccretion.hpp"
+#include "particle/formation/EnzoMethodBondiHoyleAccretion.hpp"
 #include "enzo-core/EnzoMethodBalance.hpp"
 #include "enzo-core/EnzoMethodCheck.hpp"
 #include "cosmology/EnzoMethodComovingExpansion.hpp"
 #include "cosmology/EnzoMethodCosmology.hpp"
-#include "particle/EnzoMethodDistributedFeedback.hpp"
-#include "particle/EnzoMethodFeedback.hpp"
-#include "particle/EnzoMethodFeedbackSTARSS.hpp"
-#include "particle/EnzoMethodFluxAccretion.hpp"
+#include "particle/feedback/EnzoMethodDistributedFeedback.hpp"
+#include "particle/feedback/EnzoMethodFeedback.hpp"
+#include "particle/feedback/EnzoMethodFeedbackSTARSS.hpp"
+#include "particle/formation/EnzoMethodFluxAccretion.hpp"
 #include "assorted/EnzoMethodGrackle.hpp"
 #include "gravity/EnzoMethodGravity.hpp"
 #include "assorted/EnzoMethodHeat.hpp"
 #include "hydro-mhd/EnzoMethodHydro.hpp"
-#include "particle/EnzoMethodMergeSinks.hpp"
+#include "particle/formation/EnzoMethodMergeSinks.hpp"
 #include "hydro-mhd/EnzoMethodMHDVlct.hpp"
 #include "assorted/EnzoMethodM1Closure.hpp"
 #include "gravity/EnzoMethodPmDeposit.hpp"
 #include "particle/EnzoMethodPmUpdate.hpp"
 #include "hydro-mhd/EnzoMethodPpm.hpp"
 #include "hydro-mhd/EnzoMethodPpml.hpp"
-#include "particle/EnzoMethodSinkMaker.hpp"
-#include "particle/EnzoMethodStarMaker.hpp"
-#include "particle/EnzoMethodStarMakerSTARSS.hpp"
-#include "particle/EnzoMethodStarMakerStochasticSF.hpp"
-#include "particle/EnzoMethodThresholdAccretion.hpp"
+#include "particle/formation/EnzoMethodSinkMaker.hpp"
+#include "particle/formation/EnzoMethodStarMaker.hpp"
+#include "particle/formation/EnzoMethodStarMakerSTARSS.hpp"
+#include "particle/formation/EnzoMethodStarMakerStochasticSF.hpp"
+#include "particle/formation/EnzoMethodThresholdAccretion.hpp"
 #include "assorted/EnzoMethodTurbulence.hpp"
 
 #include "gravity/EnzoMatrixDiagonal.hpp"
